% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEMAV.R
\docType{package}
\name{CEMAV}
\alias{CEMAV-package}
\alias{CEMAV}
\title{CEMAV: Cadeias Estocásticas de Memória Variável}
\description{
Pacote destinado às funções criadas ao longo da iniciação científica "Introdução às cadeias estocásticas de memória variável"
}
\author{
\strong{Maintainer}: Caio Théodore Genovese Huss Oliveira \email{caiotheodore26@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
