% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemav_bin.R
\name{cemav_bin}
\alias{cemav_bin}
\title{Cadeia estocástica de memória variável binária}
\usage{
cemav_bin(
  n,
  passado = NA,
  tam_passado = 100,
  q0 = 1/2,
  qk = function(x) 1/(x + 1),
  q_inf = 0
)
}
\arguments{
\item{n}{Número de iterações a serem geradas.}

\item{passado}{Vetor de passado da cadeia, caso deixe vazio será gerado automaticamente com p(1)=1/2.}

\item{tam_passado}{Tamanho do passado a ser gerado, caso forneça um passado, é irrelevante.}

\item{q0}{valor entre 0 e 1, P(X=1|X_(-1)=1).}

\item{qk}{função f(k) in (0,1), P(X=1|X(-1)=...=X(-k)=0, X(-k-1)=1)}

\item{q_inf}{valor entre 0 e 1, P(X=1|X(-1)=...=0)}
}
\value{
uma lista:
\itemize{
\item \code{serie} valores simulados
\item \code{regen} tempos regenerativos
\item \code{prob} probabilidade empírica de X=1
\item \code{q} valores de qk
}
}
\description{
Simula uma cadeia de memoria variavel binária com a função contexto qk
}
